libvncclient: Don't ignore screen size updates with screen.id = 0
0 is a valid screen ID. It is used by the VNC server implementation in Qemu.

As per https://github.com/rfbproto/rfbproto/blob/master/rfbproto.rst#L4017-L4022 :
" The *id* field contains an arbitrary value that the server and client
can use to map RFB screens to physical screens. The value must be
unique in the current set of screens and must be preserved for the
lifetime of that RFB screen. New ids are assigned by whichever side
creates the screen. An *id* may be reused if there has been a subsequent
update of the screen layout where the *id* was not used. "